refc: refactor tools into modules
