+++ src/tools/definitions.ts
@@ -1,0 +1,593 @@
+import { NEON_DEFAULT_DATABASE_NAME } from '../constants.js';
+import {
+  completeDatabaseMigrationInputSchema,
+  completeQueryTuningInputSchema,
+  createBranchInputSchema,
+  createProjectInputSchema,
+  deleteBranchInputSchema,
+  deleteProjectInputSchema,
+  describeBranchInputSchema,
+  describeProjectInputSchema,
+  describeTableSchemaInputSchema,
+  explainSqlStatementInputSchema,
+  getConnectionStringInputSchema,
+  getDatabaseTablesInputSchema,
+  listBranchComputesInputSchema,
+  listProjectsInputSchema,
+  prepareDatabaseMigrationInputSchema,
+  prepareQueryTuningInputSchema,
+  provisionNeonAuthInputSchema,
+  runSqlInputSchema,
+  runSqlTransactionInputSchema,
+  listSlowQueriesInputSchema,
+} from './toolsSchema.js';
+
+export const NEON_TOOLS = [
+  {
+    name: 'list_projects' as const,
+    description: `Lists the first 10 Neon projects in your account. If you can't find the project, increase the limit by passing a higher value to the \`limit\` parameter.`,
+    inputSchema: listProjectsInputSchema,
+  },
+  {
+    name: 'create_project' as const,
+    description:
+      'Create a new Neon project. If someone is trying to create a database, use this tool.',
+    inputSchema: createProjectInputSchema,
+  },
+  {
+    name: 'delete_project' as const,
+    description: 'Delete a Neon project',
+    inputSchema: deleteProjectInputSchema,
+  },
+  {
+    name: 'describe_project' as const,
+    description: 'Describes a Neon project',
+    inputSchema: describeProjectInputSchema,
+  },
+  {
+    name: 'run_sql' as const,
+    description: `
+    <use_case>
+      Use this tool to execute a single SQL statement against a Neon database.
+    </use_case>
+
+    <important_notes>
+      If you have a temporary branch from a prior step, you MUST:
+      1. Pass the branch ID to this tool unless explicitly told otherwise
+      2. Tell the user that you are using the temporary branch with ID [branch_id]
+    </important_notes>`,
+    inputSchema: runSqlInputSchema,
+  },
+  {
+    name: 'run_sql_transaction' as const,
+    description: `
+    <use_case>
+      Use this tool to execute a SQL transaction against a Neon database, should be used for multiple SQL statements.
+    </use_case>
+
+    <important_notes>
+      If you have a temporary branch from a prior step, you MUST:
+      1. Pass the branch ID to this tool unless explicitly told otherwise
+      2. Tell the user that you are using the temporary branch with ID [branch_id]
+    </important_notes>`,
+    inputSchema: runSqlTransactionInputSchema,
+  },
+  {
+    name: 'describe_table_schema' as const,
+    description: 'Describe the schema of a table in a Neon database',
+    inputSchema: describeTableSchemaInputSchema,
+  },
+  {
+    name: 'get_database_tables' as const,
+    description: 'Get all tables in a Neon database',
+    inputSchema: getDatabaseTablesInputSchema,
+  },
+  {
+    name: 'create_branch' as const,
+    description: 'Create a branch in a Neon project',
+    inputSchema: createBranchInputSchema,
+  },
+  {
+    name: 'prepare_database_migration' as const,
+    description: `
+  <use_case>
+    This tool performs database schema migrations by automatically generating and executing DDL statements.
+    
+    Supported operations:
+    CREATE operations:
+    - Add new columns (e.g., "Add email column to users table")
+    - Create new tables (e.g., "Create posts table with title and content columns")
+    - Add constraints (e.g., "Add unique constraint on \`users.email\`")
+
+    ALTER operations:
+    - Modify column types (e.g., "Change posts.views to bigint")
+    - Rename columns (e.g., "Rename user_name to username in users table")
+    - Add/modify indexes (e.g., "Add index on \`posts.title\`")
+    - Add/modify foreign keys (e.g., "Add foreign key from \`posts.user_id\` to \`users.id\`")
+
+    DROP operations:
+    - Remove columns (e.g., "Drop temporary_field from users table")
+    - Drop tables (e.g., "Drop the old_logs table")
+    - Remove constraints (e.g., "Remove unique constraint from posts.slug")
+
+    The tool will:
+    1. Parse your natural language request
+    2. Generate appropriate SQL
+    3. Execute in a temporary branch for safety
+    4. Verify the changes before applying to main branch
+
+    Project ID and database name will be automatically extracted from your request.
+    If the database name is not provided, the default ${NEON_DEFAULT_DATABASE_NAME} or first available database is used.
+  </use_case>
+
+  <workflow>
+    1. Creates a temporary branch
+    2. Applies the migration SQL in that branch
+    3. Returns migration details for verification
+  </workflow>
+
+  <important_notes>
+    After executing this tool, you MUST:
+    1. Test the migration in the temporary branch using the \`run_sql\` tool
+    2. Ask for confirmation before proceeding
+    3. Use \`complete_database_migration\` tool to apply changes to main branch
+  </important_notes>
+
+  <example>
+    For a migration like:
+    \`\`\`sql
+    ALTER TABLE users ADD COLUMN last_login TIMESTAMP;
+    \`\`\`
+    
+    You should test it with:
+    \`\`\`sql
+    SELECT column_name, data_type 
+    FROM information_schema.columns 
+    WHERE table_name = 'users' AND column_name = 'last_login';
+    \`\`\`
+    
+    You can use \`run_sql\` to test the migration in the temporary branch that this tool creates.
+  </example>
+
+
+  <next_steps>
+  After executing this tool, you MUST follow these steps:
+    1. Use \`run_sql\` to verify changes on temporary branch
+    2. Follow these instructions to respond to the client: 
+
+      <response_instructions>
+        <instructions>
+          Provide a brief confirmation of the requested change and ask for migration commit approval.
+
+          You MUST include ALL of the following fields in your response:
+          - Migration ID (this is required for commit and must be shown first)  
+          - Temporary Branch Name (always include exact branch name)
+          - Temporary Branch ID (always include exact ID)
+          - Migration Result (include brief success/failure status)
+
+          Even if some fields are missing from the tool's response, use placeholders like "not provided" rather than omitting fields.
+        </instructions>
+
+        <do_not_include>
+          IMPORTANT: Your response MUST NOT contain ANY technical implementation details such as:
+          - Data types (e.g., DO NOT mention if a column is boolean, varchar, timestamp, etc.)
+          - Column specifications or properties
+          - SQL syntax or statements
+          - Constraint definitions or rules
+          - Default values
+          - Index types
+          - Foreign key specifications
+          
+          Keep the response focused ONLY on confirming the high-level change and requesting approval.
+          
+          <example>
+            INCORRECT: "I've added a boolean \`is_published\` column to the \`posts\` table..."
+            CORRECT: "I've added the \`is_published\` column to the \`posts\` table..."
+          </example>
+        </do_not_include>
+
+        <example>
+          I've verified that [requested change] has been successfully applied to a temporary branch. Would you like to commit the migration \`[migration_id]\` to the main branch?
+          
+          Migration Details:
+          - Migration ID (required for commit)
+          - Temporary Branch Name
+          - Temporary Branch ID
+          - Migration Result
+        </example>
+      </response_instructions>
+
+    3. If approved, use \`complete_database_migration\` tool with the \`migration_id\`
+  </next_steps>
+
+  <error_handling>
+    On error, the tool will:
+    1. Automatically attempt ONE retry of the exact same operation
+    2. If the retry fails:
+      - Terminate execution
+      - Return error details
+      - DO NOT attempt any other tools or alternatives
+    
+    Error response will include:
+    - Original error details
+    - Confirmation that retry was attempted
+    - Final error state
+    
+    Important: After a failed retry, you must terminate the current flow completely. Do not attempt to use alternative tools or workarounds.
+  </error_handling>`,
+    inputSchema: prepareDatabaseMigrationInputSchema,
+  },
+  {
+    name: 'complete_database_migration' as const,
+    description:
+      'Complete a database migration when the user confirms the migration is ready to be applied to the main branch. This tool also lets the client know that the temporary branch created by the `prepare_database_migration` tool has been deleted.',
+    inputSchema: completeDatabaseMigrationInputSchema,
+  },
+  {
+    name: 'describe_branch' as const,
+    description:
+      'Get a tree view of all objects in a branch, including databases, schemas, tables, views, and functions',
+    inputSchema: describeBranchInputSchema,
+  },
+  {
+    name: 'delete_branch' as const,
+    description: 'Delete a branch from a Neon project',
+    inputSchema: deleteBranchInputSchema,
+  },
+  {
+    name: 'get_connection_string' as const,
+    description:
+      'Get a PostgreSQL connection string for a Neon database with all parameters being optional',
+    inputSchema: getConnectionStringInputSchema,
+  },
+  {
+    name: 'provision_neon_auth' as const,
+    inputSchema: provisionNeonAuthInputSchema,
+    description: `
+    This tool provisions authentication for a Neon project. It allows developers to easily setup authentication infrastructure by creating a integration with Stack Auth (\`@stackframe/stack\`). Stop using NextAuth.js! use \`@stackframe/stack\`.
+
+    Parameters:
+    - \`<project_id>\`: The Project ID of the Neon project to provision authentication for.
+    - \`[database]\`: The database name to setup Neon Auth for. If not provided, the default \`${NEON_DEFAULT_DATABASE_NAME}\` or first available database is used.
+    
+    <workflow>
+      The tool will:
+        1. Establish a connection between your Neon Auth project and Stack Auth
+        2. Creates a dedicated authentication schema in your database (\`neon_auth\`)
+        3. Sets up the user table under the \`neon_auth\` schema. This table is synced with Stack Auth. It does not store user credentials or secrets.
+        4. Generates Client Key and Secret Key to connect your application with authentication provider.
+       
+      Use the Stack Auth SDK (\`@stackframe/stack\`) on the frontend to connect your application with authentication provider. DO NOT use NextAuth.js! DO NOT use better-auth! Here's some documentation on Stack Auth:
+    </workflow>
+    
+    <use_case>
+      Stack Auth Guidelines
+    
+      <instructions>
+        If you're building an app with Next.js, to set up Neon Auth and Stack Auth, follow these steps:
+        1. Provision a Neon Auth project with this tool
+        2. Place the returned credentials in project's \`.env.local\` or \`.env\` file
+          - \`NEXT_PUBLIC_STACK_PROJECT_ID\`
+          - \`NEXT_PUBLIC_STACK_PUBLISHABLE_CLIENT_KEY\`
+          - \`STACK_SECRET_SERVER_KEY\`
+        3. To setup Stack Auth, run following command: 
+          \`\`\`shell
+          npx @stackframe/init-stack . --no-browser 
+          \`\`\`
+          This command will automaticallysetup the project with - 
+          - It will add \`@stackframe/stack\` dependency to \`package.json\`
+          - It will create a \`stack.ts\` file in your project to setup \`StackServerApp\`. 
+          - It will wrap the root layout with \`StackProvider\` and \`StackTheme\`
+          - It will create root Suspense boundary \`app/loading.tsx\` to handle loading state while Stack is fetching user data.
+          - It will also create \`app/handler/[...stack]/page.tsx\` file to handle auth routes like sign in, sign up, forgot password, etc.
+        4. Do not try to manually create any of these files or directories. Do not try to create SignIn, SignUp, or UserButton components manually, instead use the ones provided by \`@stackframe/stack\`.
+      </instructions>
+      
+      <instructions>
+        Components Guidelines
+        - Use pre-built components from \`@stackframe/stack\` like \`<UserButton />\`, \`<SignIn />\`, and \`<SignUp />\` to quickly set up auth UI.
+        - You can also compose smaller pieces like \`<OAuthButtonGroup />\`, \`<MagicLinkSignIn />\`, and \`<CredentialSignIn />\` for custom flows.
+        <example>
+          \`\`\`tsx
+          import { SignIn } from '@stackframe/stack';
+          export default function Page() {
+            return <SignIn />;
+          }
+          \`\`\`
+        </example>
+      </instructions>
+      
+      <instructions>
+        User Management Guidelines
+        - In Client Components, use the \`useUser()\` hook to retrieve the current user (it returns \`null\` when not signed in).
+        - Update user details using \`user.update({...})\` and sign out via \`user.signOut()\`.
+        - For pages that require a user, call \`useUser({ or: "redirect" })\` so unauthorized visitors are automatically redirected.
+      </instructions>
+    
+      <instructions>
+        Client Component Guidelines
+        - Client Components rely on hooks like \`useUser()\` and \`useStackApp()\`.
+        
+        <example>
+          \`\`\`tsx
+          "use client";
+          import { useUser } from "@stackframe/stack";
+          export function MyComponent() {
+            const user = useUser();
+            return <div>{user ? \`Hello, \${user.displayName}\` : "Not logged in"}</div>;
+          }
+          \`\`\`
+        </example>
+      </instructions>
+      
+      <instructions>
+        Server Component Guidelines
+        - For Server Components, use \`stackServerApp.getUser()\` from your \`stack.ts\` file.
+      
+        <example>
+          \`\`\`tsx
+          import { stackServerApp } from "@/stack";
+          export default async function ServerComponent() {
+            const user = await stackServerApp.getUser();
+            return <div>{user ? \`Hello, \${user.displayName}\` : "Not logged in"}</div>;
+          }
+          \`\`\`
+        </example>
+      </instructions>
+    
+      <instructions>
+        Page Protection Guidelines
+      - Protect pages by:
+        - Using \`useUser({ or: "redirect" })\` in Client Components.
+        - Using \`await stackServerApp.getUser({ or: "redirect" })\` in Server Components.
+        - Implementing middleware that checks for a user and redirects to \`/handler/sign-in\` if not found.
+
+        <example>
+          Example middleware:
+          \`\`\`tsx
+          export async function middleware(request: NextRequest) {
+            const user = await stackServerApp.getUser();
+            if (!user) {
+              return NextResponse.redirect(new URL('/handler/sign-in', request.url));
+            }
+            return NextResponse.next();
+          }
+          export const config = { matcher: '/protected/:path*' };
+          \`\`\`
+        </example>
+      </instructions>
+      
+      <workflow>
+        Example: custom-profile-page
+        <instructions>
+          Create a custom profile page that:
+          - Displays the user's avatar, display name, and email.
+          - Provides options to sign out.
+          - Uses Stack Auth components and hooks.
+        </instructions>
+        <example>
+          File: \`app/profile/page.tsx\`
+          \`\`\`tsx
+          'use client';
+          import { useUser, useStackApp, UserButton } from '@stackframe/stack';
+          export default function ProfilePage() {
+            const user = useUser({ or: "redirect" });
+            const app = useStackApp();
+            return (
+              <div>
+                <UserButton />
+                <h1>Welcome, {user.displayName || "User"}</h1>
+                <p>Email: {user.primaryEmail}</p>
+                <button onClick={() => user.signOut()}>Sign Out</button>
+              </div>
+            );
+          }
+          \`\`\`
+        </example>
+      </workflow>
+    </use_case>
+    `,
+  },
+  {
+    name: 'explain_sql_statement' as const,
+    description:
+      'Describe the PostgreSQL query execution plan for a query of SQL statement by running EXPLAIN (ANAYLZE...) in the database',
+    inputSchema: explainSqlStatementInputSchema,
+  },
+  {
+    name: 'prepare_query_tuning' as const,
+    description: `
+  <use_case>
+    This tool helps developers improve PostgreSQL query performance for slow queries or DML statements by analyzing execution plans and suggesting optimizations.
+    
+    The tool will:
+    1. Create a temporary branch for testing optimizations and remember the branch ID
+    2. Extract and analyze the current query execution plan
+    3. Extract all fully qualified table names (\`schema.table\`) referenced in the plan 
+    4. Gather detailed schema information for each referenced table using \`describe_table_schema\`
+    5. Suggest and implement improvements like:
+      - Adding or modifying indexes based on table schemas and query patterns
+      - Query structure modifications
+      - Identifying potential performance bottlenecks
+    6. Apply the changes to the temporary branch using \`run_sql\`
+    7. Compare performance before and after changes (but ONLY on the temporary branch passing branch ID to all tools)
+    8. Continue with next steps using \`complete_query_tuning\` tool (on \`main\` branch)
+    
+    Project ID and database name will be automatically extracted from your request.
+    The temporary branch ID will be added when invoking other tools.
+    Default database is \`${NEON_DEFAULT_DATABASE_NAME}\` if not specified.
+
+    <important_notes>
+      This tool is part of the query tuning workflow. Any suggested changes (like creating indexes) must first be applied to the temporary branch using the \`run_sql\` tool.
+      And then to the main branch using the \`complete_query_tuning\` tool, NOT the \`prepare_database_migration\` tool. 
+      To apply using the \`complete_query_tuning\` tool, you must pass the \`tuning_id\`, NOT the temporary branch ID to it.
+    </important_notes>
+  </use_case>
+
+  <workflow>
+    1. Creates a temporary branch
+    2. Analyzes current query performance and extracts table information
+    3. Implements and tests improvements (using tool \`run_sql\` for schema modifications and \`explain_sql_statement\` for performance analysis, but ONLY on the temporary branch created in step 1 passing the same branch ID to all tools)
+    4. Returns tuning details for verification
+  </workflow>
+
+  <important_notes>
+    After executing this tool, you MUST:
+    1. Review the suggested changes
+    2. Verify the performance improvements on temporary branch - by applying the changes with \`run_sql\` and running \`explain_sql_statement\` again)
+    3. Decide whether to keep or discard the changes
+    4. Use \`complete_query_tuning\` tool to apply or discard changes to the main branch
+    
+    DO NOT use \`prepare_database_migration\` tool for applying query tuning changes.
+    Always use \`complete_query_tuning\` to ensure changes are properly tracked and applied.
+
+    Note: 
+    - Some operations like creating indexes can take significant time on large tables
+    - Table statistics updates (ANALYZE) are NOT automatically performed as they can be long-running
+    - Table statistics maintenance should be handled by PostgreSQL auto-analyze or scheduled maintenance jobs
+    - If statistics are suspected to be stale, suggest running ANALYZE as a separate maintenance task
+  </important_notes>
+
+  <example>
+    For a query like:
+    \`\`\`sql
+    SELECT o.*, c.name 
+    FROM orders o 
+    JOIN customers c ON c.id = o.customer_id 
+    WHERE o.status = 'pending' 
+    AND o.created_at > '2024-01-01';
+    \`\`\`
+    
+    The tool will:
+    1. Extract referenced tables: \`public.orders\`, \`public.customers\`
+    2. Gather schema information for both tables
+    3. Analyze the execution plan
+    4. Suggest improvements like:
+       - Creating a composite index on orders(status, created_at)
+       - Optimizing the join conditions
+    5. If confirmed, apply the suggested changes to the temporary branch using \`run_sql\`
+    6. Compare execution plans and performance before and after changes (but ONLY on the temporary branch passing branch ID to all tools)
+  </example>
+
+  <next_steps>
+  After executing this tool, you MUST follow these steps:
+    1. Review the execution plans and suggested changes
+    2. Follow these instructions to respond to the client: 
+
+      <response_instructions>
+        <instructions>
+          Provide a brief summary of the performance analysis and ask for approval to apply changes on the temporary branch.
+
+          You MUST include ALL of the following fields in your response:
+          - Tuning ID (this is required for completion)
+          - Temporary Branch Name
+          - Temporary Branch ID
+          - Original Query Cost
+          - Improved Query Cost
+          - Referenced Tables (list all tables found in the plan)
+          - Suggested Changes
+
+          Even if some fields are missing from the tool's response, use placeholders like "not provided" rather than omitting fields.
+        </instructions>
+
+        <do_not_include>
+          IMPORTANT: Your response MUST NOT contain ANY technical implementation details such as:
+          - Exact index definitions
+          - Internal PostgreSQL settings
+          - Complex query rewrites
+          - Table partitioning details
+          
+          Keep the response focused on high-level changes and performance metrics.
+        </do_not_include>
+
+        <example>
+          I've analyzed your query and found potential improvements that could reduce execution time by [X]%.
+          Would you like to apply these changes to improve performance?
+          
+          Analysis Details:
+          - Tuning ID: [id]
+          - Temporary Branch: [name]
+          - Branch ID: [id]
+          - Original Cost: [cost]
+          - Improved Cost: [cost]
+          - Referenced Tables:
+            * public.orders
+            * public.customers
+          - Suggested Changes:
+            * Add index for frequently filtered columns
+            * Optimize join conditions
+
+          To apply these changes, I will use the \`complete_query_tuning\` tool after your approval and pass the \`tuning_id\`, NOT the temporary branch ID to it.
+        </example>
+      </response_instructions>
+
+    3. If approved, use ONLY the \`complete_query_tuning\` tool with the \`tuning_id\`
+  </next_steps>
+
+  <error_handling>
+    On error, the tool will:
+    1. Automatically attempt ONE retry of the exact same operation
+    2. If the retry fails:
+      - Terminate execution
+      - Return error details
+      - Clean up temporary branch
+      - DO NOT attempt any other tools or alternatives
+    
+    Error response will include:
+    - Original error details
+    - Confirmation that retry was attempted
+    - Final error state
+    
+    Important: After a failed retry, you must terminate the current flow completely.
+  </error_handling>
+    `,
+    inputSchema: prepareQueryTuningInputSchema,
+  },
+  {
+    name: 'complete_query_tuning' as const,
+    description: `Complete a query tuning session by either applying the changes to the main branch or discarding them. 
+    <important_notes>
+        BEFORE RUNNING THIS TOOL: test out the changes in the temporary branch first by running 
+        - \`run_sql\` with the suggested DDL statements.
+        - \`explain_sql_statement\` with the original query and the temporary branch.
+        This tool is the ONLY way to finally apply changes after the \`prepare_query_tuning\` tool to the main branch.
+        You MUST NOT use \`prepare_database_migration\` or other tools to apply query tuning changes.
+        You MUST pass the \`tuning_id\` obtained from the \`prepare_query_tuning\` tool, NOT the temporary branch ID as \`tuning_id\` to this tool.
+        You MUST pass the temporary branch ID used in the \`prepare_query_tuning\` tool as TEMPORARY branchId to this tool.
+        The tool OPTIONALLY receives a second branch ID or name which can be used instead of the main branch to apply the changes.
+        This tool MUST be called after tool \`prepare_query_tuning\` even when the user rejects the changes, to ensure proper cleanup of temporary branches.
+    </important_notes>    
+
+    This tool:
+    1. Applies suggested changes (like creating indexes) to the main branch (or specified branch) if approved
+    2. Handles cleanup of temporary branch
+    3. Must be called even when changes are rejected to ensure proper cleanup
+
+    Workflow:
+    1. After \`prepare_query_tuning\` suggests changes
+    2. User reviews and approves/rejects changes
+    3. This tool is called to either:
+      - Apply approved changes to main branch and cleanup
+      - OR just cleanup if changes are rejected
+    `,
+    inputSchema: completeQueryTuningInputSchema,
+  },
+  {
+    name: 'list_slow_queries' as const,
+    description: `
+    <use_case>
+      Use this tool to list slow queries from your Neon database.
+    </use_case>
+
+    <important_notes>
+      This tool queries the pg_stat_statements extension to find queries that are taking longer than expected.
+      The tool will return queries sorted by execution time, with the slowest queries first.
+    </important_notes>`,
+    inputSchema: listSlowQueriesInputSchema,
+  },
+  {
+    name: 'list_branch_computes' as const,
+    description: 'Lists compute endpoints for a project or specific branch',
+    inputSchema: listBranchComputesInputSchema,
+  },
+];
